Update crud.py
Now throws an error if any fields are missing from the request.
Due to the scope of this project, I likely wont add this check to other endpoints. This would be bad for production code, as all input validation ought to be done serverside, but it'll do for a university project.